重構OPL3寄存器寫入函數，將參數類型更改為一致的無符號字符；更新資源大小獲取函數的返回類型；調整文本顯示長度計算；禁用ESP32內存報告功能；新增音樂播放器函數聲明

+++ src/adplug/nuked/opl3.c
@@ -1486,7 +1486,7 @@ void OPL3_WriteReg(opl3_chip *chip, uint16_t reg, uint8_t v)
     }
 }
 
-void OPL3_WriteRegBuffered(int16_t reg, uint8_t v)
+void OPL3_WriteRegBuffered(uint8_t reg, uint8_t v)
 {
     uint64_t time1, time2;
     opl3_writebuf *writebuf;
