+++ src/resource.c
@@ -353,7 +353,7 @@ unsigned short RES_MKFGetChunkCount(
     return (resource_id >= Res_Count) ? 0 : g_CachedResourceIndex[resource_id].chunk_count;
 }
 
-unsigned short RES_MKFGetChunkSize(
+unsigned int RES_MKFGetChunkSize(
     unsigned int chunk_index,
     unsigned char resource_id)
 {
