+++ src/resource.h
@@ -39,7 +39,7 @@ unsigned short RES_MKFGetChunkCount(
     unsigned char resource_id);
 
 // Get the size of a specific chunk in a resource file
-unsigned short RES_MKFGetChunkSize(
+unsigned int RES_MKFGetChunkSize(
     unsigned int chunk_index,
     unsigned char resource_id);
